// mw-tools/src/game-server/Enums/FightActionType.ts
export const enum FightActionType {
	Melee = 0,
	MeleeCombo = 1,
	Magic = 2,
	MagicCombo = 3,
	Item = 4,
	Escape = 5,
	ChangePet = 6,
	Drain = 7,
	Poison = 8,
	Effect = 9,
	MultiMagic = 100,
	MultiMagicCombo = 101,
	MultiMelee = 102,
	MultiMeleeDouble = 103,
	MultiDrain = 104,
	MultiPoison = 105,
	MultiEffect = 102,
	SingleStatus = 7,
}