// mw-tools/src/game-server/Enums/FightEffect.ts
export const enum FightEffect {
	None = 0x00000000,
	Poison = 0x00000001,
	Hypnotize = 0x00000002,
	Stun = 0x00000004,
	Frailty = 0x00000008,
	Chaos = 0x00000010,
	Death = 0x00000020,
	//circle?	= 0x00000040,
	//red thingy?	= 0x00000080,
	//blue blink	= 0x00000100,
	//yellow blink	= 0x00000200,
	Enhance = 0x00000400,
	Protect = 0x00000800,
	Speed = 0x00001000,
	Repel = 0x00002000,
	Reflect = 0x00004000,
	Dead = 0x00008000,
	Leave = 0x00010000,
	Defend = 0x00020000,
}